fix(cli-mode): fix create-flow toggle and skip kandev-system wrap for passthrough
Two passthrough UX fixes flagged during testing.

1. CLI Passthrough toggle didn't move when creating a new profile in
   /settings/agents/[agentId]. The form sends snake_case patches
   (cli_passthrough) but the create flow's handleProfileChange spread them
   directly into the draft, which stores camelCase (cliPassthrough). The
   edit-profile flow already had a snake→camel translator (toAgentProfilePatch)
   — lift it to agent-save-helpers.ts and use it on both flows.

2. New tasks created with a passthrough agent had the <kandev-system>...
   MCP tool block prepended to the user's first message. Visible verbatim
   in the TUI — surprising and unhelpful. Skip the InjectKandevContext wrap
   when the agent profile / session is passthrough across all four wrap
   sites: orchestrator's StartCreatedSession + startTask, wsAddMessage on
   CREATED sessions, and the workflow auto-start handler.
