+++ apps/web/app/settings/agents/[agentId]/agent-setup-parts.tsx
@@ -20,7 +20,7 @@ import { ProfileEnvVarsSection } from "@/components/settings/agent-profile-page"
 import { CustomCLIFlagsCard } from "@/components/settings/cli-flags-field";
 import type { Agent, ModelConfig, PermissionSetting, PassthroughConfig } from "@/lib/types/http";
 import { ProfileMcpConfigCard } from "./profile-mcp-config-card";
-import type { DraftProfile, DraftAgent } from "./agent-save-helpers";
+import { toAgentProfilePatch, type DraftProfile, type DraftAgent } from "./agent-save-helpers";
 
 export type AgentHeaderProps = {
   displayName: string;
@@ -124,7 +124,7 @@ export function ProfileCardItem({
             cli_passthrough: profile.cliPassthrough ?? false,
             cli_flags: profile.cliFlags ?? [],
           }}
-          onChange={(patch) => onProfileChange(profile.id, patch)}
+          onChange={(patch) => onProfileChange(profile.id, toAgentProfilePatch(patch))}
           modelConfig={currentAgentModelConfig}
           permissionSettings={permissionSettings}
           passthroughConfig={passthroughConfig}
